+++ stdlib/Pkg3/src/API.jl
@@ -7,7 +7,7 @@ import Dates
 import LibGit2
 
 import ..depots, ..logdir, ..devdir, ..print_first_command_header
-import ..Operations, ..Display, ..GitTools
+import ..Operations, ..Display, ..GitTools, ..Pkg3
 using ..Types, ..TOML
 
 
@@ -15,8 +15,10 @@ preview_info() = printstyled("───── Preview mode ─────\n"; c
 
 include("generate.jl")
 
+parse_package(pkg) = Pkg3.REPLMode.parse_package(pkg; context=Pkg3.REPLMode.CMD_ADD)
+
 add_or_develop(pkg::Union{String, PackageSpec}; kwargs...) = add_or_develop([pkg]; kwargs...)
-add_or_develop(pkgs::Vector{String}; kwargs...)            = add_or_develop([PackageSpec(pkg) for pkg in pkgs]; kwargs...)
+add_or_develop(pkgs::Vector{String}; kwargs...)            = add_or_develop([parse_package(pkg) for pkg in pkgs]; kwargs...)
 add_or_develop(pkgs::Vector{PackageSpec}; kwargs...)       = add_or_develop(Context(), pkgs; kwargs...)
 
 function add_or_develop(ctx::Context, pkgs::Vector{PackageSpec}; mode::Symbol, kwargs...)
@@ -27,6 +29,7 @@ function add_or_develop(ctx::Context, pkgs::Vector{PackageSpec}; mode::Symbol, k
         new_git = handle_repos_develop!(ctx, pkgs)
     else
         new_git = handle_repos_add!(ctx, pkgs; upgrade_or_add=true)
+        update_registry(ctx)
     end
     project_deps_resolve!(ctx.env, pkgs)
     registry_resolve!(ctx.env, pkgs)
@@ -34,6 +37,7 @@ function add_or_develop(ctx::Context, pkgs::Vector{PackageSpec}; mode::Symbol, k
     ensure_resolved(ctx.env, pkgs, registry=true)
     Operations.add_or_develop(ctx, pkgs; new_git=new_git)
     ctx.preview && preview_info()
+    return
 end
 
 add(args...; kwargs...) = add_or_develop(args...; mode = :add, kwargs...)
@@ -53,24 +57,15 @@ function rm(ctx::Context, pkgs::Vector{PackageSpec}; kwargs...)
     manifest_resolve!(ctx.env, pkgs)
     Operations.rm(ctx, pkgs)
     ctx.preview && preview_info()
+    return
 end
 
 
-up(;kwargs...)                           = up(PackageSpec[]; kwargs...)
-up(pkg::Union{String, PackageSpec}; kwargs...)               = up([pkg]; kwargs...)
-up(pkgs::Vector{String}; kwargs...)      = up([PackageSpec(pkg) for pkg in pkgs]; kwargs...)
-up(pkgs::Vector{PackageSpec}; kwargs...) = up(Context(), pkgs; kwargs...)
-
-function up(ctx::Context, pkgs::Vector{PackageSpec};
-            level::UpgradeLevel=UPLEVEL_MAJOR, mode::PackageMode=PKGMODE_PROJECT, kwargs...)
-    print_first_command_header()
-    Context!(ctx; kwargs...)
-    ctx.preview && preview_info()
-
+function update_registry(ctx)
     # Update the registry
     errors = Tuple{String, String}[]
     if ctx.preview
-        info("Skipping updating registry in preview mode")
+        @info("Skipping updating registry in preview mode")
     else
         for reg in registries()
             if isdir(joinpath(reg, ".git"))
@@ -86,7 +81,13 @@ function up(ctx::Context, pkgs::Vector{PackageSpec};
                         return
                     end
                     branch = LibGit2.headname(repo)
-                    GitTools.fetch(repo)
+                    try
+                        GitTools.fetch(repo)
+                    catch e
+                        e isa LibGit2.GitError || rethrow(e)
+                        push!(errors, (reg, "failed to fetch from repo"))
+                        return
+                    end
                     ff_succeeded = try
                         LibGit2.merge!(repo; branch="refs/remotes/origin/$branch", fastforward=true)
                     catch e
@@ -107,7 +108,6 @@ function up(ctx::Context, pkgs::Vector{PackageSpec};
             end
         end
     end
-
     if !isempty(errors)
         warn_str = "Some registries failed to update:"
         for (reg, err) in errors
@@ -115,7 +115,20 @@ function up(ctx::Context, pkgs::Vector{PackageSpec};
         end
         @warn warn_str
     end
+    return
+end
 
+up(;kwargs...)                           = up(PackageSpec[]; kwargs...)
+up(pkg::Union{String, PackageSpec}; kwargs...)               = up([pkg]; kwargs...)
+up(pkgs::Vector{String}; kwargs...)      = up([PackageSpec(pkg) for pkg in pkgs]; kwargs...)
+up(pkgs::Vector{PackageSpec}; kwargs...) = up(Context(), pkgs; kwargs...)
+
+function up(ctx::Context, pkgs::Vector{PackageSpec};
+            level::UpgradeLevel=UPLEVEL_MAJOR, mode::PackageMode=PKGMODE_PROJECT, kwargs...)
+    print_first_command_header()
+    Context!(ctx; kwargs...)
+    ctx.preview && preview_info()
+    update_registry(ctx)
     if isempty(pkgs)
         if mode == PKGMODE_PROJECT
             for (name::String, uuidstr::String) in ctx.env.project["deps"]
@@ -135,6 +148,7 @@ function up(ctx::Context, pkgs::Vector{PackageSpec};
     end
     Operations.up(ctx, pkgs)
     ctx.preview && preview_info()
+    return
 end
 
 
@@ -149,6 +163,7 @@ function pin(ctx::Context, pkgs::Vector{PackageSpec}; kwargs...)
     project_deps_resolve!(ctx.env, pkgs)
     ensure_resolved(ctx.env, pkgs)
     Operations.pin(ctx, pkgs)
+    return
 end
 
 
@@ -163,11 +178,12 @@ function free(ctx::Context, pkgs::Vector{PackageSpec}; kwargs...)
     registry_resolve!(ctx.env, pkgs)
     ensure_resolved(ctx.env, pkgs; registry=true)
     Operations.free(ctx, pkgs)
+    return
 end
 
 
 
-test(;kwargs...)                                  = test(PackageSpec[], kwargs...)
+test(;kwargs...)                                  = test(PackageSpec[]; kwargs...)
 test(pkg::Union{String, PackageSpec}; kwargs...)  = test([pkg]; kwargs...)
 test(pkgs::Vector{String}; kwargs...)             = test([PackageSpec(pkg) for pkg in pkgs]; kwargs...)
 test(pkgs::Vector{PackageSpec}; kwargs...)        = test(Context(), pkgs; kwargs...)
@@ -186,6 +202,7 @@ function test(ctx::Context, pkgs::Vector{PackageSpec}; coverage=false, kwargs...
     manifest_resolve!(ctx.env, pkgs)
     ensure_resolved(ctx.env, pkgs)
     Operations.test(ctx, pkgs; coverage=coverage)
+    return
 end
 
 
@@ -299,6 +316,7 @@ function gc(ctx::Context=Context(); period = Dates.Week(6), kwargs...)
     byte_save_str = length(paths_to_delete) == 0 ? "" : ("saving " * @sprintf("%.3f %s", bytes, Base._mem_units[mb]))
     @info("Deleted $(length(paths_to_delete)) package installations $byte_save_str")
     ctx.preview && preview_info()
+    return
 end
 
 
@@ -317,6 +335,7 @@ function _get_deps!(ctx::Context, pkgs::Vector{PackageSpec}, uuids::Vector{UUID}
         end
         _get_deps!(ctx, pkgs, uuids)
     end
+    return
 end
 
 build(pkgs...) = build([PackageSpec(pkg) for pkg in pkgs])
@@ -349,6 +368,7 @@ function build(ctx::Context, pkgs::Vector{PackageSpec}; kwargs...)
     length(uuids) == 0 && (@info("no packages to build"); return)
     Operations.build_versions(ctx, uuids; might_need_to_resolve=true)
     ctx.preview && preview_info()
+    return
 end
 
 init() = init(Context())
@@ -357,6 +377,82 @@ function init(ctx::Context, path::String=pwd())
     print_first_command_header()
     Context!(ctx; env = EnvCache(joinpath(path, "Project.toml")))
     Operations.init(ctx)
+    return
+end
+
+#####################################
+# Backwards compatibility with Pkg2 #
+#####################################
+
+function clone(pkg::String...)
+    @warn "Pkg.clone is only kept for legacy CI script reasons, please use `add`" maxlog=1
+    add(joinpath(pkg...))
+end
+
+function dir(pkg::String, paths::String...)
+    @warn "Pkg.dir is only kept for legacy CI script reasons" maxlog=1
+    pkgid = Base.identify_package(pkg)
+    pkgid == nothing && return nothing
+    path = Base.locate_package(pkgid)
+    pkgid == nothing && return nothing
+    return joinpath(abspath(path, "..", "..", paths...))
+end
+
+
+function precompile(ctx::Context)
+    printpkgstyle(ctx, :Precompiling, "project...")
+
+    pkgids = [Base.PkgId(UUID(uuid), name) for (name, uuid) in ctx.env.project["deps"] if !(UUID(uuid) in  keys(ctx.stdlibs))]
+    if ctx.env.pkg !== nothing && isfile( joinpath( dirname(ctx.env.project_file), "src", ctx.env.pkg.name * ".jl"))
+        push!(pkgids, Base.PkgId(ctx.env.pkg.uuid, ctx.env.pkg.name))
+    end
+
+    needs_to_be_precompiled = String[]
+    for pkg in pkgids
+        paths = Base.find_all_in_cache_path(pkg)
+        sourcepath = Base.locate_package(pkg)
+        if sourcepath == nothing
+            cmderror("couldn't find path to $(pkg.name) when trying to precompilie project")
+        end
+        found_matching_precompile = false
+        for path_to_try in paths::Vector{String}
+            staledeps = Base.stale_cachefile(sourcepath, path_to_try)
+            if !(staledeps isa Bool)
+                found_matching_precompile = true
+            end
+        end
+        if !found_matching_precompile
+            # Only precompile packages that has contains `__precompile__` or `__precompile__(true)`
+            source = read(sourcepath, String)
+            if occursin(r"__precompile__\(\)|__precompile__\(true\)", source)
+                push!(needs_to_be_precompiled, pkg.name)
+            end
+        end
+    end
+
+    # Perhaps running all the imports in the same process would avoid some overheda.
+    # Julia starts pretty fast though (0.3 seconds)
+    code = join(["import " * pkg for pkg in needs_to_be_precompiled], '\n') * "\nexit(0)"
+    for (i, pkg) in enumerate(needs_to_be_precompiled)
+        code = """
+            empty!(Base.DEPOT_PATH)
+            append!(Base.DEPOT_PATH, $(repr(map(abspath, DEPOT_PATH))))
+            empty!(Base.DL_LOAD_PATH)
+            append!(Base.DL_LOAD_PATH, $(repr(map(abspath, Base.DL_LOAD_PATH))))
+            empty!(Base.LOAD_PATH)
+            append!(Base.LOAD_PATH, $(repr(Base.LOAD_PATH)))
+            import $pkg
+        """
+        printpkgstyle(ctx, :Precompiling, pkg, " [$i of $(length(needs_to_be_precompiled))]")
+        run(pipeline(ignorestatus(```
+        $(Base.julia_cmd()) -O$(Base.JLOptions().opt_level) --color=no --history-file=no
+        --startup-file=$(Base.JLOptions().startupfile != 2 ? "yes" : "no")
+        --compiled-modules="yes"
+        --depwarn=no
+        --eval $code
+        ```)))
+    end
+    return nothing
 end
 
 end # module